fix: merge pending WAL frames when reading data.db for context-window enrichment
CopilotAppDataAccess (getSessionHierarchy/getSessionContextInfo) read
~/.copilot/data.db as raw bytes via fs.readFileSync, but the Copilot app
keeps that database open in WAL mode. Recent writes -- e.g.
context_current_tokens for an actively-running CLI session -- can sit
in data.db-wal, uncommitted to the main file, for a long time. sql.js
has no concept of the WAL sidecar, so those reads silently returned a
stale snapshot missing the session's own context-window data, making
the Context Window & Long-Context Pricing section show 'No data' until
something else (the writer closing its connection) forced a checkpoint.

opencode.ts already solved this exact problem for opencode.db via a
copy + node:sqlite PRAGMA wal_checkpoint(TRUNCATE) merge. Extracted
that into a shared src/utils/sqliteWal.ts helper and applied it to
both opencode.ts and copilotAppData.ts so all sql.js-backed SQLite
reads in the extension get a consistent, checkpoint-merged snapshot.
