+++ src/opencode.ts
@@ -11,6 +11,7 @@ import type { ModelUsage, ModelId } from './types';
 import { normalizePathForComparison } from './workspaceHelpers';
 import { isUnsafeObjectKey } from './utils/protoGuard';
 import { readTextFileWithSizeGuardSync } from './utils/safeFileRead';
+import { readDbBufferWithWal, getWalMtimeMs } from './utils/sqliteWal';
 
 // Access SqlJsStatic and Database via the globally declared initSqlJs namespace.
 type SqlJsStatic = initSqlJs.SqlJsStatic;
@@ -137,81 +138,26 @@ export class OpenCodeDataAccess {
 		}
 	}
 
-	/** Returns the WAL file's mtime in milliseconds, or 0 if no WAL file exists. */
-	private getWalMtimeMs(dbPath: string): number {
-		try {
-			return fs.statSync(dbPath + '-wal').mtimeMs;
-		} catch {
-			return 0;
-		}
-	}
-
 	private isCachedDbCurrent(dbPath: string, stats: fs.Stats): boolean {
 		return this._dbCache?.path === dbPath
 			&& this._dbCache.mtimeMs === stats.mtimeMs
 			&& this._dbCache.size === stats.size
-			&& this._dbCache.walMtimeMs === this.getWalMtimeMs(dbPath);
+			&& this._dbCache.walMtimeMs === getWalMtimeMs(dbPath);
 	}
 
 	private getDbCacheKey(dbPath: string, stats: fs.Stats): string {
-		return `${dbPath}:${stats.mtimeMs}:${stats.size}:wal${this.getWalMtimeMs(dbPath)}`;
+		return `${dbPath}:${stats.mtimeMs}:${stats.size}:wal${getWalMtimeMs(dbPath)}`;
 	}
 
 	private sameDbStats(left: fs.Stats, right: fs.Stats): boolean {
 		return left.mtimeMs === right.mtimeMs && left.size === right.size;
 	}
 
-	/**
-	 * When an active WAL file is present, sql.js cannot see uncommitted WAL frames because
-	 * it reads only the raw `.db` bytes. This method copies the DB + WAL to a temp location,
-	 * opens the copy with Node's built-in SQLite (available in Node.js 22+), forces a WAL
-	 * checkpoint to merge all frames into the temp DB file, and returns the resulting buffer
-	 * so that sql.js can load a fully up-to-date snapshot.
-	 *
-	 * Returns null when no WAL is present, when the WAL is empty, or when node:sqlite is
-	 * unavailable — in all those cases the caller falls back to reading the DB file directly.
-	 */
-	private async tryReadDbWithWal(dbPath: string): Promise<Buffer | null> {
-		const walPath = dbPath + '-wal';
-		let walSize: number;
-		try {
-			walSize = fs.statSync(walPath).size;
-		} catch {
-			return null; // No WAL file — no merge needed
-		}
-		if (walSize === 0) { return null; }
-
-		try {
-			const { DatabaseSync } = require('node:sqlite') as typeof import('node:sqlite');
-			const tmpDir = path.join(os.homedir(), '.copilot', 'tmp');
-			fs.mkdirSync(tmpDir, { recursive: true, mode: 0o700 });
-			const tmpDb = path.join(tmpDir, `opencode-wal-${process.pid}-${Date.now()}-${Math.random().toString(16).slice(2)}.db`);
-			const tmpWal = tmpDb + '-wal';
-			const tmpShm = tmpDb + '-shm';
-			const shmPath = dbPath + '-shm';
-
-			fs.copyFileSync(dbPath, tmpDb);
-			fs.copyFileSync(walPath, tmpWal);
-			if (fs.existsSync(shmPath)) { fs.copyFileSync(shmPath, tmpShm); }
-
-			const nativeDb = new DatabaseSync(tmpDb);
-			nativeDb.exec('PRAGMA wal_checkpoint(TRUNCATE);');
-			nativeDb.close();
-
-			const buffer = fs.readFileSync(tmpDb);
-			for (const f of [tmpDb, tmpWal, tmpShm]) { try { fs.unlinkSync(f); } catch { /* ignore */ } }
-			return buffer;
-		} catch {
-			return null; // node:sqlite unavailable or copy failed — fall back to direct read
-		}
-	}
-
 	private async refreshOpenCodeDb(dbPath: string, stats: fs.Stats): Promise<SqlDatabase | null> {
 		let db: SqlDatabase;
 		try {
 			const SQL = await this.initSqlJs();
-			const walBuffer = await this.tryReadDbWithWal(dbPath);
-			const buffer = walBuffer ?? fs.readFileSync(dbPath);
+			const buffer = await readDbBufferWithWal(dbPath);
 			db = new SQL.Database(buffer);
 		} catch {
 			return this.getCachedDbForPath(dbPath);
@@ -227,7 +173,7 @@ export class OpenCodeDataAccess {
 		}
 
 		this.closeDbCache();
-		this._dbCache = { db, path: dbPath, mtimeMs: stats.mtimeMs, size: stats.size, walMtimeMs: this.getWalMtimeMs(dbPath) };
+		this._dbCache = { db, path: dbPath, mtimeMs: stats.mtimeMs, size: stats.size, walMtimeMs: getWalMtimeMs(dbPath) };
 		return db;
 	}
